+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2023-10-26
 
+### Added
+
+- Add `ObjectRef::contains_key`.
+
 ### Fixed
 
 - Remove data of duplicate keys when building objects.